chrom_rs::output::export

Trait Exporter

pub trait Exporter {
    type Error: Error;

    // Required methods
    fn export_single(
        &self,
        result: &SimulationResult,
        n_points: Option<usize>,
        path: &str,
    ) -> Result<(), Self::Error>;
    fn export_multi(
        &self,
        result: &SimulationResult,
        n_points: Option<usize>,
        species_names: &[&str],
        path: &str,
    ) -> Result<(), Self::Error>;
}

Abstraction trait for all export formats.

Associated type Error

Each format manages its own errors via the associated type. This avoids systematic boxing (Box<dyn Error>) and allows the caller to react precisely based on the error type.

Parameter n_points

None: exports all time steps (default behaviour)

Some(n): uniformly downsamples to n points, always guaranteeing that the first and last points are included (important to capture the end of the chromatographic peak)

Implementing this trait

A new format must implement [export_single] and [export_multi]. Formats that do not distinguish between the two cases can delegate one to the other.

Required Associated Types

type Error: Error

Error type specific to this export format.

Required Methods

fn export_single( &self, result: &SimulationResult, n_points: Option<usize>, path: &str, ) -> Result<(), Self::Error>

Exports a single-species result.

The file contains two columns: time and c_outlet.

Errors

Returns an error if:

the path is invalid or the directory does not exist

result contains no data

fn export_multi( &self, result: &SimulationResult, n_points: Option<usize>, species_names: &[&str], path: &str, ) -> Result<(), Self::Error>

Exports a multi-species result with an envelope column (c_total).

The file contains: time, c_total, then one column per species. The c_total column is the sum of all species at each time step (represents the raw detector signal, which does not distinguish species).

Errors

Returns an error if:

the number of names in species_names does not match the data

the path is invalid

result contains no data