chrom_rs::output::visualization

Module config

Plot configuration shared across visualization modules

This module defines common configuration structures used by both steady-state and chromatogram plotting functions.

Structs

PlotConfig

Configuration for customizing plots

Constants

NO_TITLE

Constant for no title (default title will be used)

Traits

IntoOptionalTitle

Helper trait to accept both String and None for optional titles